Update friends controller
